feat: added bulk and serialized inventory items

+++ database/migrations/sqlite/20260514003041_company_settings.sql
@@ -4,7 +4,7 @@ CREATE TABLE org_settings (
   id TEXT PRIMARY KEY,
   org_id TEXT NOT NULL REFERENCES orgs(id) ON DELETE CASCADE,
   currency TEXT NOT NULL,
-  vat_rate REAL NOT NULL,
+  vat_rate INTEGER NOT NULL,
   timezone TEXT NOT NULL,
   updated_at INTEGER NOT NULL DEFAULT (unixepoch()),
   created_at INTEGER NOT NULL DEFAULT (unixepoch())
